Unify euclidean-grid standalone and composable codebase
